+++ packages/flutter_best_practices_lints/lib/src/flutter_best_practices_plugin.dart
@@ -2,7 +2,9 @@
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
 // 🌎 Project imports:
+import 'package:flutter_best_practices_lints/src/rules/avoid_widget_operator_equals.dart';
 import 'package:flutter_best_practices_lints/src/rules/matching_class_and_file_name.dart';
+import 'package:flutter_best_practices_lints/src/rules/prefer_widget_class_over_widget_helper.dart';
 import 'package:flutter_best_practices_lints/src/rules/single_class_per_file.dart';
 
 /// This is the entrypoint of our custom linter
@@ -15,5 +17,7 @@ class FlutterBestPracticesPlugin extends PluginBase {
   List<LintRule> getLintRules(CustomLintConfigs configs) => <LintRule>[
     const SingleClassPerFile(),
     const MatchingClassAndFileName(),
+    const PreferWidgetClassOverWidgetHelper(),
+    const AvoidWidgetOperatorEquals(),
   ];
 }